refactor: renamed convert_gvf_to_vcf.py to convert_gvf_to_vcf_logic.py
