+++ convert_gvf_to_vcf/convert_gvf_to_vcf_logic.py
